Add an one possible way to rendering the club details.
We consider the Clubs component is a container (stateful) and the Club component is stateless.
The simplest way to display such details is just to change the state of the Clubs component (without changing the URL), but this approach has an one drawback - we cannot use a direct link to open a specific club. So we should use routing to change the URL when we go to the specific club.
However if we use a routing to display a component more likely this component will be a stateful (in our case the Club component). But in our simple way we can remain the Club component as a stateless and use this way to rendering club details.